+++ app/src/main/java/com/github/teamjcd/bpp/provider/BppAddressColumns.java
@@ -1,7 +1,7 @@
 package com.github.teamjcd.bpp.provider;
 
 public class BppAddressColumns extends BppBaseColumns {
-    public BppAddressColumns(String name, int value) {
+    public BppAddressColumns(String name, long value) {
         super(name, value);
     }
