+++ app/src/main/java/com/github/teamjcd/bpp/provider/BppDeviceClassColumns.java
@@ -1,7 +1,7 @@
 package com.github.teamjcd.bpp.provider;
 
 public class BppDeviceClassColumns extends BppBaseColumns {
-    public BppDeviceClassColumns(String name, int value) {
+    public BppDeviceClassColumns(String name, long value) {
         super(name, value);
     }
